added singleMessage view

+++ app/Http/Controllers/ContactController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\Contact;
 use App\Mail\ContactMail;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
 
 
@@ -16,7 +17,8 @@ class ContactController extends Controller
 
     public function destroy(Request $request)
     {
-        if($request['id'])
+        $isAdmin = Auth::user() && Auth::user()->admin;
+        if($isAdmin && $request['id'])
         {
             Contact::findOrFail($request['id'])->delete();
         }
@@ -26,7 +28,27 @@ class ContactController extends Controller
 
     public function show()
     {
-        return view('contactView', ["data" => Contact::all()]);
+        $isAdmin = Auth::user() && Auth::user()->admin;
+
+        if($isAdmin)
+        {
+            return view('contactView', ["data" => Contact::all()]);
+        }
+        return back()->with('error', 'You are not authorized to view messages');
+    }
+
+    public function find(Request $request)
+    {
+        $isAdmin = Auth::user() && Auth::user()->admin;
+
+        if($isAdmin)
+        {
+            $data = Contact::findOrFail($request['id']);
+             
+            return view('contactView_singleMessage', ['data'=>$data]);
+        }
+
+        return back()->with('error', 'You are not authorized to view messages');
     }
 
     public function create(Request $request)
